start  adding celery workers (#124)
* start  adding celery workers

* add celery worker

* frontend changes for new collector worker

* worker

* add osint scheduling to queue manager

* Add WorkerView to frontend

* readd tasks.py

* working RESTScheduler

* add Deprecation Warnings

* finalizing that stuff

* next_run_time to db

* fix linting error

+++ src/core/core/model/word_list.py
@@ -11,18 +11,18 @@ from core.model.acl_entry import ACLEntry, ItemType
 class WordList(BaseModel):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(), nullable=False)
-    description = db.Column(db.String(), nullable=False)
+    description = db.Column(db.String(), default=None)
     use_for_stop_words = db.Column(db.Boolean, default=False)
     link = db.Column(db.String(), nullable=True, default=None)
     entries = db.relationship("WordListEntry", cascade="all, delete-orphan")
 
-    def __init__(self, name, description="", use_for_stop_words=False, link=None, entries=None, id=None):
+    def __init__(self, name, description=None, use_for_stop_words=False, link=None, entries=None, id=None):
         self.id = id
         self.name = name
         self.description = description
         self.use_for_stop_words = use_for_stop_words
         self.link = link
-        self.entries = entries or []
+        self.entries = [WordListEntry.get(entry) for entry in entries] if entries else []
 
     @classmethod
     def find_by_name(cls, name):
@@ -66,7 +66,7 @@ class WordList(BaseModel):
             query = query.filter(
                 or_(
                     WordList.name.ilike(f"%{search}%"),
-                    WordList.description.ilike(f"%{search}%"),
+                    WordList.description.ilike(f"%{search}%"),  # type: ignore
                 )
             )
 
@@ -78,17 +78,9 @@ class WordList(BaseModel):
         items = [word_list.to_dict() for word_list in word_lists]
         return {"total_count": count, "items": items}
 
-    @classmethod
-    def load_multiple(cls, json_data: list[dict[str, Any]]) -> list["WordList"]:
-        return [cls.from_dict(data) for data in json_data]
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> "WordList":
-        return cls(**data)
-
     def to_dict(self) -> dict[str, Any]:
         data = super().to_dict()
-        data["entries"] = [entry.to_dict() for entry in self.entries]
+        data["entries"] = [entry.to_dict() for entry in self.entries if entry]
         data["tag"] = "mdi-format-list-bulleted-square"
         return data
 
@@ -124,15 +116,22 @@ class WordListEntry(BaseModel):
     @classmethod
     def delete_entries(cls, id, value):
         word_list = WordList.get(id)
+        if not word_list:
+            return "WordList not found", 404
         cls.query.filter_by(word_list_id=word_list.id).filter_by(value=value).delete()
         db.session.commit()
 
     @classmethod
     def update_word_list_entries(cls, id, entries_data):
         word_list = WordList.get(id)
+        if not word_list:
+            return "WordList not found", 404
 
         entries = cls.load_multiple(entries_data)
+        if not entries:
+            return "No entries found", 404
         for entry in entries:
             if not cls.identical(entry.value, word_list.id):
                 word_list.entries.append(entry)
                 db.session.commit()
+        return "WordList entries updated", 200
